moved all mocked computations to a different class
