+++ projects/com.nvidia.grcuda.test/src/com/nvidia/grcuda/test/runtime/stream/ComplexExecutionDAGMockTest.java
@@ -55,6 +55,8 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.executeMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.imageMockComputation;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -146,86 +148,6 @@ public class ComplexExecutionDAGMockTest {
         assertEquals(numStreams, context.getStreamManager().getNumberOfStreams());
         for (CUDAStream stream : ((GrCUDAStreamManagerMock) context.getStreamManager()).getStreams()) {
             assertTrue(context.getStreamManager().isStreamFree(stream));
-        }
-    }
-
-    // 0: K0(1c, 2) -> 2: K3(2c, 5) -> 4: K5(2c, 5c, 3) -> Repeat -> S(3)
-    //              \--------------\/
-    //              /--------------/\
-    // 1: K1(3c, 4) -> 3: K4(4c, 6) -> 5: K6(4c, 6c, 1) -> Repeat -> S(1)
-    public static List<GrCUDAComputationalElement> hitsMockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a1 = new DeviceArrayMock(10);
-        DeviceArrayMock a2 = new DeviceArrayMock(10);
-        DeviceArrayMock a3 = new DeviceArrayMock(10);
-        DeviceArrayMock a4 = new DeviceArrayMock(10);
-        DeviceArrayMock a5 = new DeviceArrayMock(1);
-        DeviceArrayMock a6 = new DeviceArrayMock(1);
-        List<GrCUDAComputationalElement> computations = new ArrayList<>();
-        int numIterations = 2;
-        for (int i = 0; i < numIterations; i++) {
-            // hub1 -> auth2
-            computations.add(new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a1, true), new ArgumentMock(a2))));
-            // auth1 -> hub2
-            computations.add(new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a3, true), new ArgumentMock(a4))));
-            // auth2 -> auth_norm
-            computations.add(new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a2, true), new ArgumentMock(a5))));
-            // hub2 -> hub_norm
-            computations.add(new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a4, true), new ArgumentMock(a6))));
-            // auth2, auth_norm -> auth1
-            computations.add(new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a2, true), new ArgumentMock(a5, true), new ArgumentMock(a3))));
-            // hub2, hub_norm -> hub1
-            computations.add(new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a4, true), new ArgumentMock(a6, true), new ArgumentMock(a1))));
-        }
-        computations.add(new SyncExecutionMock(context, Collections.singletonList(new ArgumentMock(a3))));
-        computations.add(new SyncExecutionMock(context, Collections.singletonList(new ArgumentMock(a1))));
-        return computations;
-    }
-
-    // 0: B1(1c, 2) -> 3: S1(2c, 5) -------------------------------------------------------------> 10: C(10c, 2c, 5c, 11) -> X
-    // 1: B2(1c, 3) -> 4: S2(3c, 6) -------------------------------------> 9: C(9c, 3c, 6c, 10) /
-    // 2: B3(1c, 4) -> 5: E1(4c, 7) -> 7: E3(4, 7, 8) -> 8: U(1c, 4, 9) /
-    //             \-> 6: E2(4c, 8) /
-    public static List<GrCUDAComputationalElement> imageMockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a1 = new DeviceArrayMock(10);
-        DeviceArrayMock a2 = new DeviceArrayMock(10);
-        DeviceArrayMock a3 = new DeviceArrayMock(10);
-        DeviceArrayMock a4 = new DeviceArrayMock(10);
-        DeviceArrayMock a5 = new DeviceArrayMock(10);
-        DeviceArrayMock a6 = new DeviceArrayMock(10);
-        DeviceArrayMock a7 = new DeviceArrayMock(1);
-        DeviceArrayMock a8 = new DeviceArrayMock(1);
-        DeviceArrayMock a9 = new DeviceArrayMock(10);
-        DeviceArrayMock a10 = new DeviceArrayMock(10);
-        DeviceArrayMock a11 = new DeviceArrayMock(10);
-        return Arrays.asList(
-            // blur
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a1, true), new ArgumentMock(a2))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a1, true), new ArgumentMock(a3))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a1, true), new ArgumentMock(a4))),
-            // sobel
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a2, true), new ArgumentMock(a5))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a3, true), new ArgumentMock(a6))),
-            // extend
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a4, true), new ArgumentMock(a7))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a4, true), new ArgumentMock(a8))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a4), new ArgumentMock(a7), new ArgumentMock(a8))),
-            // unsharpen
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a1, true), new ArgumentMock(a4), new ArgumentMock(a9))),
-            // combine
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a9, true), new ArgumentMock(a3, true),
-                    new ArgumentMock(a6, true), new ArgumentMock(a10))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a10, true), new ArgumentMock(a2, true),
-                    new ArgumentMock(a5, true), new ArgumentMock(a11))),
-            new SyncExecutionMock(context, Collections.singletonList(new ArgumentMock(a11)))
-        );
-    }
-
-    /**
-     * Schedule for execution a sequence of mock GrCUDAComputationalElement;
-     */
-    public static void executeMockComputation(List<GrCUDAComputationalElement> computations) throws UnsupportedTypeException {
-        for (GrCUDAComputationalElement c : computations) {
-            c.schedule();
         }
     }
 }
