+++ projects/com.nvidia.grcuda.test/src/com/nvidia/grcuda/test/runtime/stream/MultiGPUExecutionDAGMockTest.java
@@ -21,6 +21,17 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.executeMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.executeMockComputationAndValidate;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.forkJoinMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.hitsMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.imageMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.joinPipeline2MockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.joinPipeline3MockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.joinPipeline4MockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.joinPipelineMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.manyIndependentKernelsMockComputation;
+import static com.nvidia.grcuda.test.util.mock.GrCUDAComputationsMock.manyKernelsMockComputation;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
@@ -40,171 +51,6 @@ public class MultiGPUExecutionDAGMockTest {
         this.retrieveNewStreamPolicy = retrieveNewStreamPolicy;
     }
 
-    /**
-     * Schedule for execution a sequence of mock GrCUDAComputationalElement,
-     * and validate that the GPU scheduling of the computation is the one expected.
-     * @param computations a sequence of computations to be scheduled
-     * @param gpuScheduling a list of gpu identifiers. Each identifier "i" represents the GPU scheduling for the i-th computation;
-     * @throws UnsupportedTypeException
-     */
-    public static void executeMockComputationAndValidate(List<GrCUDAComputationalElement> computations, List<Integer> gpuScheduling) throws UnsupportedTypeException {
-        assertEquals(computations.size(), gpuScheduling.size());
-        for (int i = 0; i < computations.size(); i++) {
-            GrCUDAComputationalElement c = computations.get(i);
-            c.schedule();
-            int expected = gpuScheduling.get(i);
-            int actual = c.getStream().getStreamDeviceId();
-            if (expected != actual) {
-                System.out.println("wrong GPU allocation for kernel " + i + "=" + c + "; expected=" + expected + "; actual=" + actual);
-            }
-            assertEquals(expected, actual);
-        }
-    }
-
-    // K0 -> K4 -> K8 ---> K10
-    // K1 -> K5 /     \--> K11
-    // K2 -> K6 -> K9 -\-> K12
-    // K3 -> K7 /------\-> K13
-    public static List<GrCUDAComputationalElement> manyKernelsMockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a = new DeviceArrayMock(10);
-        DeviceArrayMock b = new DeviceArrayMock(10);
-        DeviceArrayMock c = new DeviceArrayMock(10);
-        DeviceArrayMock d = new DeviceArrayMock(10);
-        return Arrays.asList(
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(b))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(c))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(d))),
-
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(b))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(c))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(d))),
-
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(b))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(c), new ArgumentMock(d))),
-
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a, true))),
-            // When using stream-aware and 4 GPUs, this is scheduled on device 2 (of 4) as device 1 has synced the computation on it (with K8),
-            // and device 2 is the first device with fewer streams;
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a, true), new ArgumentMock(b))),
-            // When using stream-aware and 4 GPUs, this is scheduled on device 3 (reuse the stream of K9);
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a, true), new ArgumentMock(c))),
-            // When using stream-aware and 4 GPUs, this is scheduled on device 2 (device with fewer streams, device 1 has 2);
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a, true), new ArgumentMock(d)))
-        );
-    }
-
-    // Simply schedule 10 kernels on independent data;
-    public static List<GrCUDAComputationalElement> manyIndependentKernelsMockComputation(AsyncGrCUDAExecutionContext context) {
-        return Arrays.asList(
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10)))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(new DeviceArrayMock(10))))
-        );
-    }
-
-    // (Ar) --> (A, B) --> (A, B, C) -> (A, B, D)
-    // (Br) -/         /             /
-    // (Cr) ----------/             /
-    // (Dr) -----------------------/
-    public static List<GrCUDAComputationalElement> joinPipelineMockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a = new DeviceArrayMock(10);
-        DeviceArrayMock b = new DeviceArrayMock(10);
-        DeviceArrayMock c = new DeviceArrayMock(10);
-        DeviceArrayMock d = new DeviceArrayMock(100);
-        return Arrays.asList(
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(b, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(c, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(d, true))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(b))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(b), new ArgumentMock(c))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(b), new ArgumentMock(d)))
-        );
-    }
-
-    // (Ar) --> (Ar, B) --> (A, B, C)
-    // (Br) -/           /
-    // (Cr) --> (C, D) -/
-    // (Dr) -/
-    public static List<GrCUDAComputationalElement> joinPipeline2MockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a = new DeviceArrayMock(10);
-        DeviceArrayMock b = new DeviceArrayMock(10);
-        DeviceArrayMock c = new DeviceArrayMock(10);
-        DeviceArrayMock d = new DeviceArrayMock(100);
-        return Arrays.asList(
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(b, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(c, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(d, true))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a, true), new ArgumentMock(b))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(c), new ArgumentMock(d))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(b), new ArgumentMock(c)))
-        );
-    }
-
-    // (Ar) --> (Ar, B) --> (B, C)
-    // (Br) -/           /
-    // (Cr) --> (C, D) -/
-    // (Dr) -/
-    public static List<GrCUDAComputationalElement> joinPipeline3MockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a = new DeviceArrayMock(10);
-        DeviceArrayMock b = new DeviceArrayMock(10);
-        DeviceArrayMock c = new DeviceArrayMock(10);
-        DeviceArrayMock d = new DeviceArrayMock(100);
-        return Arrays.asList(
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a, false))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(b, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(c, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(d, true))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a, true), new ArgumentMock(b))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(c), new ArgumentMock(d))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(b), new ArgumentMock(c)))
-        );
-    }
-
-    // (Ar) --> (Ar, B) -> (A, C, D) -> (A, C)
-    // (Br) -/          /
-    // (Cr) -----------/
-    // (Dr) ----------/
-    public static List<GrCUDAComputationalElement> joinPipeline4MockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a = new DeviceArrayMock(10);
-        DeviceArrayMock b = new DeviceArrayMock(10);
-        DeviceArrayMock c = new DeviceArrayMock(10);
-        DeviceArrayMock d = new DeviceArrayMock(100);
-        return Arrays.asList(
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(b, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(c, true))),
-                new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(d, true))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a, true), new ArgumentMock(b))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(c), new ArgumentMock(d))),
-                new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a), new ArgumentMock(c)))
-        );
-    }
-
-    // (X) --> (Z) --> (A)
-    // (Y) -/      \-> (B)
-    public static List<GrCUDAComputationalElement> forkJoinMockComputation(AsyncGrCUDAExecutionContext context) {
-        DeviceArrayMock a1 = new DeviceArrayMock(10);
-        DeviceArrayMock a2 = new DeviceArrayMock(10);
-        return Arrays.asList(
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a1))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a2))),
-            new KernelExecutionMock(context, Arrays.asList(new ArgumentMock(a1), new ArgumentMock(a2))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a1))),
-            new KernelExecutionMock(context, Collections.singletonList(new ArgumentMock(a2)))
-        );
-    }
-
     private final static int IMAGE_NUM_STREAMS = 4;
     private final static int HITS_NUM_STREAMS = 2;
 
@@ -218,7 +64,7 @@ public class MultiGPUExecutionDAGMockTest {
                     .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.SINGLE_GPU)
                     .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                     .setNumberOfGPUsToUse(i).setNumberOfAvailableGPUs(i).build();
-            ComplexExecutionDAGMockTest.executeMockComputation(ComplexExecutionDAGMockTest.imageMockComputation(context));
+            executeMockComputation(imageMockComputation(context));
             context.getDeviceList().forEach(d -> d.getStreams().forEach(s -> assertEquals(0, s.getStreamDeviceId())));
             assertEquals(IMAGE_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
             assertEquals(IMAGE_NUM_STREAMS, context.getStreamManager().getDevice(0).getNumberOfFreeStreams());
@@ -236,7 +82,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.STREAM_AWARE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(1).setNumberOfAvailableGPUs(1).build();
-        ComplexExecutionDAGMockTest.executeMockComputation(ComplexExecutionDAGMockTest.imageMockComputation(context));
+        executeMockComputation(imageMockComputation(context));
         context.getDeviceList().forEach(d -> d.getStreams().forEach(s -> assertEquals(0, s.getStreamDeviceId())));
         assertEquals(IMAGE_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
         assertEquals(IMAGE_NUM_STREAMS, context.getStreamManager().getDevice(0).getNumberOfFreeStreams());
@@ -254,7 +100,7 @@ public class MultiGPUExecutionDAGMockTest {
                     .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.SINGLE_GPU)
                     .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                     .setNumberOfGPUsToUse(i).setNumberOfAvailableGPUs(i).build();
-            ComplexExecutionDAGMockTest.executeMockComputation(ComplexExecutionDAGMockTest.hitsMockComputation(context));
+            executeMockComputation(hitsMockComputation(context));
             context.getDeviceList().forEach(d -> d.getStreams().forEach(s -> assertEquals(0, s.getStreamDeviceId())));
             assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
             assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getDevice(0).getNumberOfFreeStreams());
@@ -272,7 +118,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.STREAM_AWARE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(1).setNumberOfAvailableGPUs(1).build();
-        ComplexExecutionDAGMockTest.executeMockComputation(ComplexExecutionDAGMockTest.hitsMockComputation(context));
+        executeMockComputation(hitsMockComputation(context));
         context.getDeviceList().forEach(d -> d.getStreams().forEach(s -> assertEquals(0, s.getStreamDeviceId())));
         assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
         assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getDevice(0).getNumberOfFreeStreams());
@@ -293,7 +139,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.STREAM_AWARE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.imageMockComputation(context),
+        executeMockComputationAndValidate(imageMockComputation(context),
                 Arrays.asList(
                         0, 1, 0,
                         0, 1,
@@ -311,7 +157,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.STREAM_AWARE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.imageMockComputation(context),
+        executeMockComputationAndValidate(imageMockComputation(context),
                 Arrays.asList(
                         0, 1, 2,
                         0, 1,
@@ -329,7 +175,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.STREAM_AWARE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.hitsMockComputation(context),
+        executeMockComputationAndValidate(hitsMockComputation(context),
                 Arrays.asList(0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0));
         assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
     }
@@ -344,7 +190,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
         // Same as 2 GPUs, it never makes sense to use the 3rd GPU;
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.hitsMockComputation(context),
+        executeMockComputationAndValidate(hitsMockComputation(context),
                 Arrays.asList(0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0));
         assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
     }
@@ -408,7 +254,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.ROUND_ROBIN)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.imageMockComputation(context),
+        executeMockComputationAndValidate(imageMockComputation(context),
                 Arrays.asList(
                         0, 1, 2,
                         0, 1,
@@ -426,7 +272,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.ROUND_ROBIN)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.hitsMockComputation(context),
+        executeMockComputationAndValidate(hitsMockComputation(context),
                 Arrays.asList(0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0));
         assertEquals(HITS_NUM_STREAMS, context.getStreamManager().getNumberOfStreams());
     }
@@ -570,7 +416,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.MIN_TRANSFER_SIZE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.imageMockComputation(context),
+        executeMockComputationAndValidate(imageMockComputation(context),
                 Arrays.asList(
                         0, 0, 0,
                         0, 0,
@@ -590,7 +436,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
         // After the first iteration, GPU0 has a3 up to date, GPU1 has a4. So GPU0 is chosen as it comes first,
         // and the scheduling collapses to GPU0;
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.hitsMockComputation(context),
+        executeMockComputationAndValidate(hitsMockComputation(context),
                 Arrays.asList(0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0));
         assertEquals(3, context.getStreamManager().getNumberOfStreams());
     }
@@ -705,7 +551,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.MIN_TRANSFER_SIZE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.imageMockComputation(context),
+        executeMockComputationAndValidate(imageMockComputation(context),
                 Arrays.asList(
                         0, 0, 0,
                         0, 0,
@@ -723,7 +569,7 @@ public class MultiGPUExecutionDAGMockTest {
                 .setDeviceSelectionPolicy(DeviceSelectionPolicyEnum.MIN_TRANSFER_SIZE)
                 .setDependencyPolicy(DependencyPolicyEnum.WITH_CONST)
                 .setNumberOfGPUsToUse(numGPU).setNumberOfAvailableGPUs(numGPU).build();
-        executeMockComputationAndValidate(ComplexExecutionDAGMockTest.hitsMockComputation(context),
+        executeMockComputationAndValidate(hitsMockComputation(context),
                 Arrays.asList(0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 0));
         assertEquals(2, context.getStreamManager().getNumberOfStreams());
     }
